refactor(cmd/Stringer): move XXStringer types to own file
- renames types from `peer` and `ref` to `peerStringer` &`refStringer`, and moves them  to `cmd/Stringer.go`
- adds `logStringer` and `String` formatting func
- adds output to pager for log
- moves `printItems` and `fmtItem` from `cmd/peers.go` to `cmd/print.go`
