+++ cmd/peers.go
@@ -1,12 +1,8 @@
 package cmd
 
 import (
-	"bytes"
 	"encoding/json"
 	"fmt"
-	"io"
-	"os"
-	"os/exec"
 
 	util "github.com/datatogether/api/apiutil"
 	"github.com/ghodss/yaml"
@@ -212,29 +208,6 @@ func (o *PeersOptions) Info() (err error) {
 	return
 }
 
-type peer config.ProfilePod
-
-func (p peer) String() string {
-	w := &bytes.Buffer{}
-	if p.Online {
-		fmt.Fprintf(w, "%s | %s\n", p.Peername, "online")
-	} else {
-		fmt.Fprintf(w, "%s\n", p.Peername)
-	}
-	fmt.Fprintf(w, "profile ID: %s\n", p.ID)
-	if len(p.NetworkAddrs) > 0 {
-		fmt.Fprintf(w, "address:    %s\n", p.NetworkAddrs[0])
-	}
-	fmt.Fprintln(w, "")
-	return w.String()
-}
-
-type stringer string
-
-func (s stringer) String() string {
-	return string(s) + "\n"
-}
-
 // List shows a list of peers
 func (o *PeersOptions) List() (err error) {
 
@@ -273,51 +246,11 @@ func (o *PeersOptions) List() (err error) {
 
 		items = make([]fmt.Stringer, len(res))
 		for i, p := range res {
-			items[i] = peer(*p)
+			items[i] = peerStringer(*p)
 		}
 	}
 
 	return printItems(o.Out, items)
-}
-
-func printItems(w io.Writer, items []fmt.Stringer) error {
-	// TODO (ramfox): This is POSIX specific, need to expand!
-	envPager := os.Getenv("PAGER")
-	if envPager == "" {
-		envPager = "less"
-	}
-
-	buf := &bytes.Buffer{}
-	pager := exec.Command(envPager)
-	pager.Stdin = buf
-	pager.Stdout = w
-
-	prefix := []byte("    ")
-	for i, item := range items {
-		buf.WriteString(fmtItem(i+1, item.String(), prefix))
-	}
-
-	return pager.Run()
-}
-
-func fmtItem(i int, item string, prefix []byte) string {
-	var res []byte
-	bol := true
-	b := []byte(item)
-	d := []byte(fmt.Sprintf("%d", i))
-	prefix1 := append(d, prefix[len(d):]...)
-	for i, c := range b {
-		if bol && c != '\n' {
-			if i == 0 {
-				res = append(res, prefix1...)
-			} else {
-				res = append(res, prefix...)
-			}
-		}
-		res = append(res, c)
-		bol = c == '\n'
-	}
-	return string(res)
 }
 
 // Connect attempts to connect to a peer
